+++ packages/backend/src/nest/registration/registration.service.ts
@@ -4,35 +4,82 @@ import { extractPendingCsrs, issueCertificate } from './registration.functions'
 import { ErrorCodes, ErrorMessages, PermsData, RegisterOwnerCertificatePayload, SocketActionTypes } from '@quiet/types'
 import { RegistrationEvents } from './registration.types'
 import Logger from '../common/logger'
+import { StorageService } from '../storage/storage.service'
 
 @Injectable()
 export class RegistrationService extends EventEmitter implements OnModuleInit {
   private readonly logger = Logger(RegistrationService.name)
-  public certificates: string[] = []
-  private _permsData: PermsData
+  private permsData: PermsData
+  private storageService: StorageService
+  private registrationEvents: { csrs: string[] }[] = []
+  private registrationEventInProgress = false
 
   constructor() {
     super()
   }
 
   onModuleInit() {
-    this.on(
-      RegistrationEvents.REGISTER_USER_CERTIFICATE,
-      async (payload: { csrs: string[]; certificates: string[]; id: string }) => {
-        await this.issueCertificates(payload)
-      }
-    )
+    this.on(RegistrationEvents.REGISTER_USER_CERTIFICATE, async (payload: { csrs: string[] }) => {
+      // Save the registration event and then try to process it, but
+      // since we only process a single event at a time, it might not
+      // get processed until other events have been processed.
+      this.registrationEvents.push(payload)
+      await this.tryIssueCertificates()
+    })
   }
 
-  private async issueCertificates(payload: { csrs: string[]; certificates: string[]; id?: string }) {
+  public init(storageService: StorageService) {
+    this.storageService = storageService
+  }
+
+  public setPermsData(permsData: PermsData) {
+    this.permsData = permsData
+  }
+
+  public async tryIssueCertificates() {
+    this.logger('Trying to issue certificates', this.registrationEventInProgress, this.registrationEvents)
+    // Process only a single registration event at a time so that we
+    // do not register two certificates with the same name.
+    if (!this.registrationEventInProgress) {
+      // Get the next event.
+      const event = this.registrationEvents.shift()
+      if (event) {
+        this.logger('Issuing certificates', event)
+        // Event processing in progress
+        this.registrationEventInProgress = true
+
+        // Await the processing function and make sure everything that
+        // needs to be done in order is awaited inside this function.
+        await this.issueCertificates({
+          ...event,
+          certificates: (await this.storageService?.loadAllCertificates()) as string[],
+        })
+
+        this.logger('Finished issuing certificates')
+        // Event processing finished
+        this.registrationEventInProgress = false
+
+        // Re-run this function if there are more events to process
+        if (this.registrationEvents.length > 0) {
+          setTimeout(this.tryIssueCertificates.bind(this), 0)
+        }
+      }
+    }
+  }
+
+  private async issueCertificates(payload: { csrs: string[]; certificates: string[] }) {
     // Lack of permsData means that we are not the owner of the
     // community in the official model of the app, however anyone can
     // modify the source code, put malicious permsData here, issue
     // false certificates and try to trick other users. To prevent
     // that, peers verify that anything that is written to the
     // certificate store is signed by the owner.
-    if (!this._permsData) {
-      if (payload.id) this.emit(RegistrationEvents.FINISHED_ISSUING_CERTIFICATES_FOR_ID, { id: payload.id })
+    //
+    // NOTE: There may be a race condition here if we try to issue
+    // certs before permsData is set. We may want to refactor this or
+    // add the ability to retry.
+    if (!this.permsData) {
+      this.logger('Not issuing certificates due to missing perms data')
       return
     }
 
@@ -44,20 +91,11 @@ export class RegistrationService extends EventEmitter implements OnModuleInit {
         await this.registerUserCertificate(csr)
       })
     )
-
-    if (payload.id) this.emit(RegistrationEvents.FINISHED_ISSUING_CERTIFICATES_FOR_ID, { id: payload.id })
-  }
-
-  public set permsData(perms: PermsData) {
-    this._permsData = {
-      certificate: perms.certificate,
-      privKey: perms.privKey,
-    }
   }
 
   public async registerOwnerCertificate(payload: RegisterOwnerCertificatePayload): Promise<void> {
-    this._permsData = payload.permsData
-    const result = await issueCertificate(payload.userCsr.userCsr, this._permsData)
+    this.permsData = payload.permsData
+    const result = await issueCertificate(payload.userCsr.userCsr, this.permsData)
     if (result?.cert) {
       this.emit(SocketActionTypes.SAVED_OWNER_CERTIFICATE, {
         communityId: payload.communityId,
@@ -74,10 +112,15 @@ export class RegistrationService extends EventEmitter implements OnModuleInit {
   }
 
   public async registerUserCertificate(csr: string): Promise<void> {
-    const result = await issueCertificate(csr, this._permsData)
+    const result = await issueCertificate(csr, this.permsData)
     this.logger('DuplicatedCertBug', { result })
     if (result?.cert) {
-      this.emit(RegistrationEvents.NEW_USER, { certificate: result.cert })
+      // Save certificate (awaited) so that we are sure that the certs
+      // are saved before processing the next round of CSRs.
+      // Otherwise, we could issue a duplicate certificate.
+
+      // @ts-ignore
+      await this.storageService?.saveCertificate({ certificate: result.cert })
     }
   }
 }